# digital/chunks to symbols: be less inefficient