--- a/gr-digital/python/digital/bindings/chunks_to_symbols_python.cc
+++ b/gr-digital/python/digital/bindings/chunks_to_symbols_python.cc
@@ -14,7 +14,7 @@
 /* BINDTOOL_GEN_AUTOMATIC(0)                                                       */
 /* BINDTOOL_USE_PYGCCXML(0)                                                        */
 /* BINDTOOL_HEADER_FILE(chunks_to_symbols.h)                                        */
-/* BINDTOOL_HEADER_FILE_HASH(92d3e85da4622d93c34d8226cbaaecf0)                     */
+/* BINDTOOL_HEADER_FILE_HASH(a7e4da5deb77580b19fd40a649ee11a8)                     */
 /***********************************************************************************/
 
 #include <pybind11/complex.h>