The block formerly took the tags from `get_tags_in_range` (which are sorted),
gave them to `tag_checker`, which sorted them, and then went through every
input sample, checking its index against the next tag.

Removed the tag_checker; that saves us a sorting of a sorted vector per work call.

The other optimization is to not dispatch any tag encountered to the own
message handler; instead, the handler is called directly.

Found #4898 while doing this.