@marcusmueller 2021-07-16
hand-wavy Benchmarking:

![benchmark_chunks_to_symbols](https://user-images.githubusercontent.com/958972/125997364-7f398b73-2e95-4299-8d8f-b61ac1401e4e.png)

Vector source affinity `[1]`, C_t_S affinity `[0]`

State  |  Rate of indexes per second | Speedup over baseline
-------|--------------------------------------|--
Before| 9.28·10⁷ | 100%
After | 43.2·107 | 466%