@marcusmueller 2021-07-16
@willcode um, you added the backport tag, but a backport will require you to undo the public header changes (which boil down to using const and unsigned where applicable); I didn't think you'd want to backport this, sorry :(