@willcode 2021-07-16
Changing int to unsigned where the value already had to be unsigned should be compatible, right?

The const changes are private.