@willcode 2021-07-16
@marcusmueller Given that the only user of D() is probably Python, which doesn't care ... if you think someone will use -Wsign-conversion -Werror, then we can keep the API the same and cast internally. Or, if you really don't want this backported at all, that's also fine - let me know.