@marcusmueller 2021-07-16
> Changing int to unsigned where the value already had to be unsigned should be compatible, right?

It does change the function signature in a way that the mangling changes, so it's at least an ABI break. 

If anyone inherits from `chunk_to_symbol`, changing of signedness, at it means the function signatures aren't compatible anymore, would also be breach of API.

I don't mind this being backported, but I do think the signed->unsigned thing warrants me patching the commit.