@willcode 2021-07-16
I always thought of the mangled signature as ABI. Will have to read up on opinions. Practically, it's likely harmless here, but I agree we should try to do the "right thing".