@marcusmueller 2021-07-17
> I always thought of the mangled signature as ABI.

And I'd agree. 

But the thing is: in C++, you can't override a `unsigned int foobar(int)`  method with a `int foobar(unsigned int)` function, or use them in the same `std::function` type and so on; separate types!